src/microservices/pythonAGR/requirements.txt
mysql-connector-python
python-dotenv
pandas
fastapi
uvicorn
